Move fileContext logic into one file.
