+++ src/index.tsx
@@ -4,14 +4,17 @@ import "./index.css";
 import App from "./app";
 import { ThemeProvider } from "./themeContext";
 import { OutlineProvider } from "./outlineContext";
+import { FileProvider } from "./fileContext";
 
 ReactDOM.render(
-  // Website wrapped in ThemeProvider so dark and light mode is
-  // available anywhere in the app
-  <OutlineProvider>
-    <ThemeProvider>
-      <App />
-    </ThemeProvider>
-  </OutlineProvider>,
+  // App wrapped in various providers that allows use of their
+  // contexts by any component in the app.
+  <FileProvider>
+    <OutlineProvider>
+      <ThemeProvider>
+        <App />
+      </ThemeProvider>
+    </OutlineProvider>
+  </FileProvider>,
   document.getElementById("root")
 );
